requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
idna==3.3
psycopg2-binary==2.9.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
